server/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
pydantic==2.5.0
sqlalchemy==2.0.23
alembic==1.12.1
python-dotenv==1.0.0
openai==1.3.7
httpx==0.25.2
pytest==7.4.3
pytest-asyncio==0.21.1
pymilvus
chromadb
langchain-huggingface==0.0.3
langsmith==0.1.94
motor
requests
mammoth
motor